Merge 3.3.7

+++ src/mongo/db/storage/mmap_v1/record_access_tracker.cpp
@@ -36,8 +36,8 @@
 #include "mongo/config.h"
 #include "mongo/db/storage/mmap_v1/record.h"
 #include "mongo/platform/bits.h"
-#include "mongo/util/concurrency/threadlocal.h"
 #include "mongo/util/clock_source.h"
+#include "mongo/util/concurrency/threadlocal.h"
 #include "mongo/util/debug_util.h"
 #include "mongo/util/net/listen.h"
 #include "mongo/util/processinfo.h"
@@ -231,7 +231,7 @@ bool RecordAccessTracker::Rolling::access(size_t region,
     if (rarelyCount++ % (2048 / BigHashSize) == 0) {
         Date_t now = cs->now();
 
-        if (now - _lastRotate > Seconds(RotateTimeSecs)) {
+        if (now - _lastRotate > Seconds(static_cast<int64_t>(RotateTimeSecs))) {
             _rotate(cs);
         }
     }
